pyproject.toml
[tool.poetry]
name = "DE School 2024 Final Project"
version = "0.1.0"
description = "Final project of Quantori data enginnering school 2024"
authors = ["Dmitry Patrushev <impda@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.lint.dependencies]
pre-commit = "^3.7.1"
isort = "^5.13.2"
ruff = "^0.4.10"

[tool.poetry.group.dev.dependencies]
apache-airflow-providers-postgres = "^5.11.2"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.isort]
known-third-party = ["airflow", "sqlalchemy"]
known-first-party = ["utils"]
